Reorganize files
Reorganized files to follow go conventions
