+++ .github/workflows/build.yml
@@ -18,4 +18,4 @@ jobs:
           go-version: '1.24.x'
 
       - name: Build
-        run: go build -v ./src/...
+        run: go build -v .
