+++ .github/workflows/test.yml
@@ -18,4 +18,7 @@ jobs:
           go-version: '1.24.x'
 
       - name: Test
-        run: go test -v ./test/...
+        run: |
+          go test -coverprofile=coverage.out -v ./...
+          TOTAL=$(go tool cover -func=coverage.out | grep total | awk '{print substr($3, 1, length($3)-1)}')
+          echo "Total coverage: $TOTAL%"
