+++ .gitignore
@@ -1,2 +1,1 @@
-*.db
-/build/**
+/.build
