+++ cfg/windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package config
+package cfg
 
 import (
 	"os"
